Fixing the types of size properties

+++ models/src/main/java/com/vimeo/networking2/VideoSourceFile.kt
@@ -69,7 +69,7 @@ data class VideoSourceFile(
      * The file size of the video.
      */
     @Json(name = "size")
-    val size: Int? = null,
+    val size: Long? = null,
 
     /**
      * The source link for the video file.
